requirements.txt
attrs==21.2.0
autopep8==1.6.0
aws-cdk.assets==1.130.0
aws-cdk.aws-apigateway==1.130.0
aws-cdk.aws-applicationautoscaling==1.130.0
aws-cdk.aws-autoscaling==1.130.0
aws-cdk.aws-autoscaling-common==1.130.0
aws-cdk.aws-autoscaling-hooktargets==1.130.0
aws-cdk.aws-certificatemanager==1.130.0
aws-cdk.aws-cloudformation==1.130.0
aws-cdk.aws-cloudfront==1.130.0
aws-cdk.aws-cloudwatch==1.130.0
aws-cdk.aws-codebuild==1.130.0
aws-cdk.aws-codecommit==1.130.0
aws-cdk.aws-codedeploy==1.130.0
aws-cdk.aws-codeguruprofiler==1.130.0
aws-cdk.aws-codepipeline==1.130.0
aws-cdk.aws-codepipeline-actions==1.130.0
aws-cdk.aws-codestarnotifications==1.130.0
aws-cdk.aws-cognito==1.130.0
aws-cdk.aws-ec2==1.130.0
aws-cdk.aws-ecr==1.130.0
aws-cdk.aws-ecr-assets==1.130.0
aws-cdk.aws-ecs==1.130.0
aws-cdk.aws-efs==1.130.0
aws-cdk.aws-elasticloadbalancing==1.130.0
aws-cdk.aws-elasticloadbalancingv2==1.130.0
aws-cdk.aws-events==1.130.0
aws-cdk.aws-events-targets==1.130.0
aws-cdk.aws-globalaccelerator==1.130.0
aws-cdk.aws-iam==1.130.0
aws-cdk.aws-kinesis==1.130.0
aws-cdk.aws-kinesisfirehose==1.130.0
aws-cdk.aws-kms==1.130.0
aws-cdk.aws-lambda==1.130.0
aws-cdk.aws-logs==1.130.0
aws-cdk.aws-route53==1.130.0
aws-cdk.aws-route53-targets==1.130.0
aws-cdk.aws-s3==1.130.0
aws-cdk.aws-s3-assets==1.130.0
aws-cdk.aws-sam==1.130.0
aws-cdk.aws-secretsmanager==1.130.0
aws-cdk.aws-servicediscovery==1.130.0
aws-cdk.aws-signer==1.130.0
aws-cdk.aws-sns==1.130.0
aws-cdk.aws-sns-subscriptions==1.130.0
aws-cdk.aws-sqs==1.130.0
aws-cdk.aws-ssm==1.130.0
aws-cdk.aws-stepfunctions==1.130.0
aws-cdk.cloud-assembly-schema==1.130.0
aws-cdk.core==1.130.0
aws-cdk.custom-resources==1.130.0
aws-cdk.cx-api==1.130.0
aws-cdk.pipelines==1.130.0
aws-cdk.region-info==1.130.0
boto3==1.19.8
botocore==1.22.8
cattrs==1.8.0
constructs==3.3.161
jmespath==0.10.0
jsii==1.41.0
publication==0.0.3
pycodestyle==2.8.0
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
